+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## v1.2.0
+#### 🚀 Features
+
+- LLM server unit tests (PR: #90)
+- Implement chat templates (PR: #92)
+- Stop chat functionality (PR: #95)
+- Keep only the llamafile binary (PR: #97)
+
+#### 🐛 Fixes
+
+- Fix remote server functionality (PR: #96)
+- Fix Max issue needing to run llamafile manually the first time (PR: #98)
+
+#### 📦 General
+
+- Async startup support (PR: #89)
+
+
 ## v1.1.1
 #### 📦 General